% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk_county.R
\name{xwalk_county}
\alias{xwalk_county}
\title{Collapse County-Level Data to New Baseline Borders}
\usage{
xwalk_county(data, y0, yb, merge_key = c("state", "county"),
  var_value = NULL, var_count = NULL)
}
\arguments{
\item{data}{county level data (likely from ICPSR 2896)}

\item{y0}{initial year of data}

\item{yb}{baseline year to transform data towards}

\item{merge_key}{vector identifying the ICPSR state and county code
if using ICPSR 2896, this is just state and county
like so: c("state", "county")}

\item{var_value}{variables that are values like average temperature or average wage
as a vector like c("avgtemp", "avgwage")}

\item{var_count}{variables that are counts like population or number of farms
as a vector like c("totpop", "urb25")}
}
\description{
Take county data, identify variables of interest, standardize borders to baseline
}
